# apps/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Applad API
  version: 1.0.0
  description: Self-hosted Backend-as-a-Service API
  license:
    name: BSD-3-Clause
servers:
  - url: /v1
    description: Default (relative)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-Applad-Key
  parameters:
    projectHeader:
      name: X-Applad-Project
      in: header
      required: true
      schema:
        type: string
...
paths:
...
  /projects:
    get:
      summary: List projects
      tags: [Projects]
      responses:
        '200':
          description: Project list
    post:
      summary: Create project
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Created
  /projects/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get project
      tags: [Projects]
      responses:
        '200':
          description: Project details
    patch:
      summary: Update project
      tags: [Projects]
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete project
      tags: [Projects]
      responses:
        '204':
          description: Deleted
  /projects/{projectId}/keys:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: List API keys
      tags: [Projects]
      responses:
        '200':
          description: Key list
    post:
      summary: Create API key
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Key created (secret returned only once)
...
  /regions/{regionId}:
    get:
      summary: Get region details
      tags: [Regions]
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Region details
  /regions/project:
    get:
      summary: List regions assigned to the project
      tags: [Regions]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      responses:
        '200':
          description: Project regions
    post:
      summary: Assign a region to the project
      tags: [Regions]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [regionId]
              properties:
                regionId:
                  type: string
      responses:
        '201':
          description: Region assigned

  # ── Webhooks ───────────────────────────────────────────────────────────────
...